fix: brought back connector.js file
